delete files instead of marking for deletion

+++ mwx-core/src/main/java/ca/encodeous/mwx/core/utils/Utils.java
@@ -503,10 +503,14 @@ public class Utils {
                 if(sub.isDirectory()){
                     stk.push(sub);
                 }else{
-                    sub.deleteOnExit();
+                    if(!sub.delete()){
+                        sub.deleteOnExit();
+                    }
                 }
             }
-            f.deleteOnExit();
+            if(!f.delete()){
+                f.deleteOnExit();
+            }
         }
     }
 }
